OK except Split & strnstr

+++ ft_strlen.c
@@ -6,13 +6,13 @@
 /*   By: lumaret <lumaret@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/08 15:50:39 by lumaret           #+#    #+#             */
-/*   Updated: 2023/11/29 17:28:55 by lumaret          ###   ########.fr       */
+/*   Updated: 2023/11/30 14:02:29 by lumaret          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *str)
+size_t	ft_strlen(const char *str)
 {
 	int	i;
